querydsl 도입 querydsl로 리팩토링

+++ library-app/src/test/kotlin/com/group/libraryapp/service/book/BookServiceTest.kt
@@ -139,13 +139,13 @@ class BookServiceTest @Autowired constructor(
 
         //then
         assertThat(results).hasSize(3)
-        assertCount(results, BookType.COMPUTER, 2)
-        assertCount(results, BookType.LANGUAGE, 3)
-        assertCount(results, BookType.ECONOMY, 1)
+        assertCount(results, BookType.COMPUTER, 2L)
+        assertCount(results, BookType.LANGUAGE, 3L)
+        assertCount(results, BookType.ECONOMY, 1L)
 
     }
 
-    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Int) {
+    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Long) {
         assertThat(results.first { result -> result.type == type }.count).isEqualTo(count)
     }
